Support both raw IO and filesystem modes.
In raw IO mode, we copy the data file from the host to address 0 of the
NVMe device and then p2pdma it. However, often the data is already in
the NVMe device and it has some filesystem. In this case, we shouldn't
copy the file (again) to the NVMe device. We can simply p2pdma it.

Filesystem mode is the default.

Signed-off-by: Martin Ichilevici de Oliveira <martin.i.oliveira@gmail.com>